feat: Add empty state to table and viewport components
Also wrap the table and viewport components.
